Owner reference for extension webhooks
